proximity-service/bs/requirements.txt
Flask
flask-sqlalchemy
flask-migrate
python-dotenv
pymysql
